// generated/github.ts
export enum ContributionLevel {
  /** Lowest 25% of days of contributions. */
  FirstQuartile = 'FIRST_QUARTILE',
  /** Highest 25% of days of contributions. More contributions than the third quartile. */
  FourthQuartile = 'FOURTH_QUARTILE',
  /** No contributions occurred. */
  None = 'NONE',
  /** Second lowest 25% of days of contributions. More contributions than the first quartile. */
  SecondQuartile = 'SECOND_QUARTILE',
  /** Second highest 25% of days of contributions. More contributions than second quartile, less than the fourth quartile. */
  ThirdQuartile = 'THIRD_QUARTILE'
}
// ...
export type GitHubUserQuery = { user?: { login: string, name?: string | null, bio?: string | null, avatarUrl: any, url: any, company?: string | null, location?: string | null, followers: { totalCount: number }, following: { totalCount: number }, contributionsCollection: { contributionYears: Array<number>, totalCommitContributions: number, totalIssueContributions: number, totalPullRequestContributions: number, contributionCalendar: { totalContributions: number, weeks: Array<{ contributionDays: Array<{ contributionLevel: ContributionLevel, contributionCount: number }> }> } }, repositories: { totalCount: number, nodes?: Array<{ name: string, description?: string | null, url: any, stargazers: { totalCount: number }, watchers: { totalCount: number }, forks: { totalCount: number } } | null> | null } } | null };